[issue-#51] Rename BeanDefinitionReader -> BeanDefinitionScanner
