--- src/test/java/com/bringframework/reader/TestService.java
@@ -1,4 +1,0 @@
-package com.bringframework.reader;
-
-public class TestService {
-}
