+++ src/test/java/com/bringframework/scanner/TestService.java
@@ -1,0 +1,4 @@
+package com.bringframework.scanner;
+
+public class TestService {
+}
